+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [2.2.0] - 2024-11-05
 
 ### Added
 
@@ -66,7 +66,8 @@
 
 - Initial release.
 
-[Unreleased]: https://github.com/thechnet/adamatic/compare/v2.1.0...HEAD
+[Unreleased]: https://github.com/thechnet/adamatic/compare/v2.2.0...HEAD
+[2.2.0]: https://github.com/thechnet/adamatic/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/thechnet/adamatic/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/thechnet/adamatic/compare/v1.3.0...v2.0.0
 [1.3.0]: https://github.com/thechnet/adamatic/compare/v1.2.0...v1.3.0